Avoid killing same process twice in macOS
When you spawned a process and you kills it, the child processes of it
will also be killed automatically on macOS.

Tests kill -INT child processes of spawned processes, which makes the
spawned processes stop in normal scenarios.

However, if a spawned process is timed out, it will be killed
automatically but the child processes of it won't be killed on Linux
(and probably on Windows).