Update to fit with a fix in 2.0.0-canary.18
In @11ty/eleventy@2.0.0-canary.18, a dependency monitoring issue was
fixed, which was Eleventy had been using platform dependent paths only
in watchers.
( 11ty/eleventy#2479 )

Now, if eleventy-sass detects the fix in Eleventy, it uses posix paths
all of the places, otherwise uses them most of the places but uses
platform dependent paths in dependencies in the dependencyMap as before.